+++ ewp-api-agent/src/main/java/kr/co/ewp/api/model/ChargingDischargingScheduleItemModel.java
@@ -7,8 +7,8 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ChargingDischargingScheduleItemModel { /*** 12.12 이우람 추가 ***/
   private Date timestamp;// timestamp start timestamp
-  private Integer scheduledCEnergy;// target for charging (Wh)
-  private Integer scheduledDEnergy;// target for discharging (Wh)
+  private Float scheduledCEnergy;// target for charging (Wh)
+  private Float scheduledDEnergy;// target for discharging (Wh)
 
   public Date getTimestamp() {
     return timestamp;
@@ -18,19 +18,19 @@ public class ChargingDischargingScheduleItemModel { /*** 12.12 이우람 추가 
     this.timestamp = timestamp;
   }
 
-  public Integer getScheduledCEnergy() {
+  public Float getScheduledCEnergy() {
   	return scheduledCEnergy;
   }
   
-  public void setScheduledCEnergy(Integer scheduledCEnergy) {
+  public void setScheduledCEnergy(Float scheduledCEnergy) {
   	this.scheduledCEnergy = scheduledCEnergy;
   }
   
-  public Integer getScheduledDEnergy() {
+  public Float getScheduledDEnergy() {
   	return scheduledDEnergy;
   }
   
-  public void setScheduledDEnergy(Integer scheduledDEnergy) {
+  public void setScheduledDEnergy(Float scheduledDEnergy) {
   	this.scheduledDEnergy = scheduledDEnergy;
   }
